RedEdit: Fix build

+++ src/apps/resedit/BitmapView.cpp
@@ -561,7 +561,7 @@ BitmapView::BitmapFromClipboard(void)
 			be_clipboard->Unlock();
 			
 			const void *buffer;
-			int32 bufferLength;
+			ssize_t bufferLength;
 			
 			BRect frame;
 			color_space cspace = B_NO_COLOR_SPACE;
@@ -591,7 +591,7 @@ BitmapView::BitmapFromClipboard(void)
 		
 		case CLIP_BEOS: {
 			const void *buffer;
-			int32 bufferLength;
+			ssize_t bufferLength;
 			
 			BRect frame;
 			color_space cspace = B_NO_COLOR_SPACE;
